requirements.txt
<<<<<<< HEAD
# Runtime dependencies
requests>=2.28
huggingface-hub>=0.14.1
GitPython>=3.1.31
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Dev / test dependencies
pytest>=7.0
pytest-cov>=4.0
coverage>=7.2
flake8>=6.0
mypy>=1.0
isort>=5.12
=======
# Runtime dependencies
requests>=2.28
huggingface-hub>=0.14.1
GitPython>=3.1.31
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Dev / test dependencies
pytest>=7.0
pytest-cov>=4.0
coverage>=7.2
flake8>=6.0
mypy>=1.0
isort>=5.12
>>>>>>> 807d2a987c68d9a2a6b6a9d9221946bb882393d3
